refactor(db): rediseño de entidades base, inventario físico y seeder inicial
- Fix: Relación bidireccional entre Warehouse (1) y Location (N).
- Feat: Creación de PurchaseOrderItem para detallar las órdenes de compra.
- Refactor: Separación lógica de Product (catálogo) y StockLevel (existencias físicas).
- Feat: Configuración de PostgreSQL y adición de DataSeeder para ambiente de desarrollo.
- Feat: Mapeo de campos SAP (sapCode, sapCenterCode) en entidades clave.

+++ src/main/java/com/visco/backend/models/entities/StockLevel.java
@@ -1,8 +1,8 @@
 package com.visco.backend.models.entities;
 
 import jakarta.persistence.*;
+import java.math.BigDecimal;
 import lombok.Data;
-
 
 @Data
 @Entity
@@ -22,8 +22,8 @@ public class StockLevel {
   private Location location;
 
   @Column(nullable = false)
-  private Double currentStock;
+  private BigDecimal currentStock;
 
   @Column(nullable = false)
-  private Double pendingStock;
+  private BigDecimal pendingStock;
 }
